back-end/requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg2==2.9.9
sqlparse==0.4.4
typing_extensions==4.11.0
certifi==2023.7.22
charset-normalizer==3.2.0
django-cors-headers==4.2.0
idna==3.4
oauthlib==3.2.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
urllib3==2.0.4
